Reactivate regression tests and bump PHP to 8.1
