+++ .github/workflows/generate_phar.yml
@@ -17,7 +17,7 @@ jobs:
 
     strategy:
       matrix:
-        php: [ '8.0' ]
+        php: [ '8.1' ]
         dependency-version: [ 'prefer-stable' ]
 
     name: Release phar
@@ -52,7 +52,7 @@ jobs:
         run: ant package -D-phar:filename=./phpmd.phar && ./phpmd.phar --version
 
       - name: Sign phar
-        if: github.repository == 'phpmd/phpmd'
+        if: github.repository == 'phpmd/phpmd' && github.event_name == 'release'
         env:
           PASSPHRASE: ${{ secrets.PASSPHRASE }}
           SECRET_KEY: ${{ secrets.SECRET_KEY }}
