+++ .scrutinizer.yml
@@ -1,7 +1,7 @@
 build:
   environment:
     php:
-      version: 8.0.0
+      version: 8.1.0
 
 checks:
   php:
