+++ composer.json
@@ -36,18 +36,18 @@
   },
   "minimum-stability": "stable",
   "require": {
-    "php": "^8.0.0",
+    "php": "^8.1",
     "ext-xml": "*",
     "composer/xdebug-handler": "^3.0",
     "pdepend/pdepend": "^2.16.1"
   },
   "require-dev": {
-    "ext-simplexml": "*",
-    "ext-json": "*",
     "easy-doc/easy-doc": "^1.3.2",
+    "ext-json": "*",
+    "ext-simplexml": "*",
     "gregwar/rst": "^1.0",
     "mikey179/vfsstream": "^1.6.8",
-    "phpunit/phpunit": "^9.6.15 || ^10.5.2",
+    "phpunit/phpunit": "^10.5.20",
     "squizlabs/php_codesniffer": "^3.8.0"
   },
   "autoload": {
@@ -63,5 +63,8 @@
     "cs-check": "phpcs -p --runtime-set ignore_errors_on_exit 1 --runtime-set ignore_warnings_on_exit 1",
     "cs-fix": "phpcbf -p --runtime-set ignore_errors_on_exit 1 --runtime-set ignore_warnings_on_exit 1",
     "build-website": "easy-doc build src/site/config.php --verbose"
+  },
+  "config": {
+    "sort-packages": true
   }
 }
