+++ src/test/php/PHPMD/Regression/StaticVariablesFlaggedAsUnusedTicket020RegressionTest.php
@@ -20,18 +20,16 @@ namespace PHPMD\Regression;
 use PHPMD\Rule\UnusedLocalVariable;
 
 /**
- * Regression test for issue 019.
- *
- * @covers \stdClass
+ * Regression test for issue 020.
  */
-class SuperGlobalsFlaggedAsUnusedTicket019RegressionTestCase extends AbstractRegressionTestCase
+class StaticVariablesFlaggedAsUnusedTicket020RegressionTest extends AbstractRegressionTestCase
 {
     /**
      * testRuleDoesNotApplyToAnySuperGlobalVariable
      *
      * @return void
      */
-    public function testRuleDoesNotApplyToAnySuperGlobalVariable()
+    public function testRuleDoesNotApplyToAnyStaticLocalVariable()
     {
         $rule = new UnusedLocalVariable();
         $rule->setReport($this->getReportWithNoViolation());
