+++ src/test/php/PHPMD/Regression/UnusedParameterArgvTicket14990109RegressionTest.php
@@ -25,10 +25,8 @@ use PHPMD\RuleSet;
  *
  * @link https://www.pivotaltracker.com/story/show/14990109
  * @since 1.1.0
- *
- * @covers \stdClass
  */
-class UnusedParameterArgvTicket14990109RegressionTestCase extends AbstractRegressionTestCase
+class UnusedParameterArgvTicket14990109RegressionTest extends AbstractRegressionTestCase
 {
     /**
      * testRuleDoesNotApplyToFunctionParameterNamedArgv
