+++ src/test/resources/files/Regression/ExcessivePublicCountRuleNeverExecutedTicket015Regression/testRuleSetInvokesRuleForClassInstance.php
