+++ src/test/resources/files/Regression/InvalidUnusedLocalVariableAndFormalParameterTicket007Regression/testFormalParameterUsedInDoubleQuoteStringGetsNotReported.php
