+++ src/test/resources/files/Regression/InvalidUnusedLocalVariableAndFormalParameterTicket007Regression/testLocalVariableUsedInDoubleQuoteStringGetsNotReported.php
