+++ src/test/resources/files/Regression/MaximumNestingLevelTicket24975295Regression/testLocalVariableUsedInDoubleQuoteStringGetsNotReported.php
