+++ src/test/resources/files/Regression/StaticVariablesFlaggedAsUnusedTicket020Regression/testRuleDoesNotApplyToAnyStaticLocalVariable.php
