+++ src/test/resources/files/Regression/SuperGlobalsFlaggedAsUnusedTicket019Regression/testRuleDoesNotApplyToAnySuperGlobalVariable.php
