+++ src/test/resources/files/Regression/SuppressWarningsNotAppliesToUnusedPrivateMethod036Regression/testRuleDoesNotApplyToPrivateMethodWithSuppressWarningsAnnotation.php
